Created definitions for data structures
