+++ docs/data.md
@@ -1,0 +1,9 @@
+# Data
+
+Tudo relacionado a dados vai ficar em `$lib/data`, isso inclui:
+
+- Definições de tipos em `types.ts`
+- Mocks em `mocks/`
+- Funções de fetch para carregar dados do backend em `fetch/`
+
+Se um tipo de dados não precisa de métodos especiais, apenas crie como um objeto
