+++ src/lib/data/mocks/matches.ts
@@ -1,11 +1,4 @@
-export type Match = {
-    id: number, 
-    home: string, 
-    visitor: string, 
-    score: number[],
-    started_at: Date,
-    finished: Date | undefined
-}
+import type { Match } from "../types";
 
 export const data: Match[] = [
     {
