+++ src/lib/data/types.ts
@@ -1,0 +1,8 @@
+export type Match = {
+    id: number, 
+    home: string, 
+    visitor: string, 
+    score: number[],
+    started_at: Date,
+    finished: Date | undefined
+}
